fix: 네비게이션 바 아이콘 변경
