+++ frontend/src/components/Icons/HomeIcon.tsx
@@ -11,14 +11,14 @@ export default function HomeIcon(props: HomeIconProps) {
 
   return (
     <svg
-      width="24"
-      height="24"
-      viewBox="0 0 20 20"
+      width="27"
+      height="27"
+      viewBox="0 0 27 27"
       fill="none"
       xmlns="http://www.w3.org/2000/svg"
     >
       <path
-        d="M19.444 7.47307L17.774 17.4731C17.6613 18.1687 17.3072 18.8024 16.774 19.2631C16.2058 19.7161 15.5007 19.9628 14.774 19.9631H5.16402C4.46134 19.9569 3.78312 19.7043 3.24766 19.2492C2.71219 18.7941 2.35346 18.1656 2.23402 17.4731L0.564023 7.47307C0.473584 6.8961 0.553519 6.30527 0.794023 5.77307C1.03445 5.24361 1.42356 4.79544 1.91402 4.48307L8.39402 0.48307C8.86746 0.191034 9.41276 0.036377 9.96902 0.036377C10.5253 0.036377 11.0706 0.191034 11.544 0.48307L18.014 4.48307C18.521 4.78307 18.925 5.23007 19.174 5.76307C19.4301 6.29415 19.524 6.88892 19.444 7.47307Z"
+        d="M23.625 12.6564V22.7814C23.625 23.0052 23.5361 23.2198 23.3779 23.3781C23.2196 23.5363 23.005 23.6252 22.7812 23.6252H16.875C16.6512 23.6252 16.4366 23.5363 16.2784 23.3781C16.1201 23.2198 16.0312 23.0052 16.0312 22.7814V17.2971C16.0312 17.1852 15.9868 17.0779 15.9077 16.9988C15.8286 16.9196 15.7213 16.8752 15.6094 16.8752H11.3906C11.2787 16.8752 11.1714 16.9196 11.0923 16.9988C11.0132 17.0779 10.9688 17.1852 10.9688 17.2971V22.7814C10.9688 23.0052 10.8799 23.2198 10.7216 23.3781C10.5634 23.5363 10.3488 23.6252 10.125 23.6252H4.21875C3.99497 23.6252 3.78036 23.5363 3.62213 23.3781C3.4639 23.2198 3.375 23.0052 3.375 22.7814V12.6564C3.37521 12.209 3.55313 11.7799 3.86965 11.4636L12.3071 3.02609C12.6236 2.70986 13.0526 2.53223 13.5 2.53223C13.9474 2.53223 14.3764 2.70986 14.6929 3.02609L23.1304 11.4636C23.4469 11.7799 23.6248 12.209 23.625 12.6564Z"
         fill={
           isActive ? theme.colorPalette.grey[500] : theme.colorPalette.grey[300]
         }
