+++ frontend/src/components/Icons/RouletteItemIcon.tsx
@@ -1,0 +1,40 @@
+import { useTheme } from '@emotion/react';
+
+interface ScissorsIconProps {
+  isActive: boolean;
+}
+
+export default function RouletteItemIcon(props: ScissorsIconProps) {
+  const { isActive } = props;
+
+  const theme = useTheme();
+
+  return (
+    <svg
+      width="48"
+      height="19"
+      viewBox="0 0 48 19"
+      fill="none"
+      xmlns="http://www.w3.org/2000/svg"
+    >
+      <path
+        d="M5.85914 3.92L12.9191 3.96C13.4125 3.96 13.7525 4.02 13.9391 4.14C14.2325 4.32667 14.3791 4.89333 14.3791 5.84C14.3791 6.32 14.2925 6.73333 14.1191 7.08C13.9591 7.41333 13.5391 8.26 12.8591 9.62C12.1925 10.98 11.7258 11.92 11.4591 12.44C10.6325 14.1067 9.75247 15.7267 8.81914 17.3C8.63247 17.6067 8.45247 17.8133 8.27914 17.92C8.11914 18.0267 7.85247 18.08 7.47914 18.08C4.66581 18.08 2.77247 18.04 1.79914 17.96C1.14581 17.9067 0.819141 17.6533 0.819141 17.2C0.819141 16.9467 0.925807 16.6267 1.13914 16.24C2.29914 14.1733 3.30581 12.4333 4.15914 11.02C5.02581 9.59333 5.55914 8.70667 5.75914 8.36C5.95914 8.01333 6.05914 7.78667 6.05914 7.68C6.05914 7.4 5.90581 7.26 5.59914 7.26C5.59914 7.26 4.31247 7.28667 1.73914 7.34C1.45914 7.34 1.25247 7.22 1.11914 6.98C0.985807 6.72667 0.919141 6.42667 0.919141 6.08C0.919141 5.72 0.939141 5.35333 0.979141 4.98C1.01914 4.59333 1.12581 4.33333 1.29914 4.2C1.47247 4.06667 1.94581 3.98667 2.71914 3.96C3.50581 3.93333 4.55247 3.92 5.85914 3.92Z"
+        fill={
+          isActive ? theme.colorPalette.grey[500] : theme.colorPalette.grey[300]
+        }
+      />
+      <path
+        d="M21.9489 0.399999L30.7739 0.449999C31.3906 0.449999 31.8156 0.524999 32.0489 0.675C32.4156 0.908333 32.5989 1.61667 32.5989 2.8C32.5989 3.4 32.4906 3.91667 32.2739 4.35C32.0739 4.76667 31.5489 5.825 30.6989 7.525C29.8656 9.225 29.2823 10.4 28.9489 11.05C27.9156 13.1333 26.8156 15.1583 25.6489 17.125C25.4156 17.5083 25.1906 17.7667 24.9739 17.9C24.7739 18.0333 24.4406 18.1 23.9739 18.1C20.4573 18.1 18.0906 18.05 16.8739 17.95C16.0573 17.8833 15.6489 17.5667 15.6489 17C15.6489 16.6833 15.7823 16.2833 16.0489 15.8C17.4989 13.2167 18.7573 11.0417 19.8239 9.275C20.9073 7.49167 21.5739 6.38333 21.8239 5.95C22.0739 5.51667 22.1989 5.23333 22.1989 5.1C22.1989 4.75 22.0073 4.575 21.6239 4.575C21.6239 4.575 20.0156 4.60833 16.7989 4.675C16.4489 4.675 16.1906 4.525 16.0239 4.225C15.8573 3.90833 15.7739 3.53333 15.7739 3.1C15.7739 2.65 15.7989 2.19167 15.8489 1.725C15.8989 1.24167 16.0323 0.916666 16.2489 0.749999C16.4656 0.583332 17.0573 0.483332 18.0239 0.449999C19.0073 0.416666 20.3156 0.399999 21.9489 0.399999Z"
+        fill={
+          isActive ? theme.colorPalette.grey[500] : theme.colorPalette.grey[300]
+        }
+      />
+      <path
+        d="M38.8591 3.92L45.9191 3.96C46.4125 3.96 46.7525 4.02 46.9391 4.14C47.2325 4.32667 47.3791 4.89333 47.3791 5.84C47.3791 6.32 47.2925 6.73333 47.1191 7.08C46.9591 7.41333 46.5391 8.26 45.8591 9.62C45.1925 10.98 44.7258 11.92 44.4591 12.44C43.6325 14.1067 42.7525 15.7267 41.8191 17.3C41.6325 17.6067 41.4525 17.8133 41.2791 17.92C41.1191 18.0267 40.8525 18.08 40.4791 18.08C37.6658 18.08 35.7725 18.04 34.7991 17.96C34.1458 17.9067 33.8191 17.6533 33.8191 17.2C33.8191 16.9467 33.9258 16.6267 34.1391 16.24C35.2991 14.1733 36.3058 12.4333 37.1591 11.02C38.0258 9.59333 38.5591 8.70667 38.7591 8.36C38.9591 8.01333 39.0591 7.78667 39.0591 7.68C39.0591 7.4 38.9058 7.26 38.5991 7.26C38.5991 7.26 37.3125 7.28667 34.7391 7.34C34.4591 7.34 34.2525 7.22 34.1191 6.98C33.9858 6.72667 33.9191 6.42667 33.9191 6.08C33.9191 5.72 33.9391 5.35333 33.9791 4.98C34.0191 4.59333 34.1258 4.33333 34.2991 4.2C34.4725 4.06667 34.9458 3.98667 35.7191 3.96C36.5058 3.93333 37.5525 3.92 38.8591 3.92Z"
+        fill={
+          isActive ? theme.colorPalette.grey[500] : theme.colorPalette.grey[300]
+        }
+      />
+    </svg>
+  );
+}
